Refactor userCard into its own component. Update UserList and UserList.test to accomodate these changes.  Update UserCard Test.
Add color check tests back to UserCard.test.js
